requirements.txt
# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0

# Machine Learning - Classical
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Deep Learning - CNN (NEW)
tensorflow==2.13.0
keras==2.13.1

# Progress bar
tqdm==4.66.1

# Dataset download
duckduckgo-search==3.9.6
requests==2.31.0

# Visualization
matplotlib==3.8.2
scikit-image==0.22.0
